# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/accounts/admin/{tenant}/clients": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": "get client list",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "client"
                ],
                "summary": "new client",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "{",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/clients/default": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": "get client",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/clients/{clientId}": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": "get client",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "clientId",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "client"
                ],
                "summary": "update client",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "clientId",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "client"
                ],
                "summary": "delete client",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "clientId",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/clients/{clientId}/redirect-uris": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": "get client redirect uris",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "clientId",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "client"
                ],
                "summary": "new client redirect uri",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "clientId",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/clients/{clientId}/redirect-uris/{uriId}": {
            "post": {
                "tags": [
                    "client"
                ],
                "summary": "new client redirect uri",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "clientId",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "client"
                ],
                "summary": "delete client redirect uris",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "clientId",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "uriId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/clients/{clientId}/secret/{secretId}": {
            "delete": {
                "tags": [
                    "client"
                ],
                "summary": "delete client secret",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "clientId",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "secretId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/clients/{clientId}/secrets": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": "get client secrets",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "clientId",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "client"
                ],
                "summary": "new client secret",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "clientId",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}